+++ lib/yarn.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require 'json'
+
+def yarn_package_installed?(_package)
+  return false unless File.exist?('package.json')
+
+  json = JSON.parse(File.read('package.json'))
+  json['dependencies'] && json['dependencies']['bootstrap'].present?
+end