+++ recipes.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 apply_recipe 'webpacker'
-apply_recipe 'rails_ujs_webpacker'
 apply_recipe 'rails_overwrites'
 # NOTE: bootstrap recipe should be invoked before simple_form to use bootstrap layout when
 #       installing simple_form
@@ -18,11 +17,9 @@ apply_recipe 'factory_bot_rails'
 apply_recipe 'rspec_rails'
 apply_recipe 'database_cleaner' # must run after rspec_rails
 apply_recipe 'rubocop'
-apply_recipe 'simple_form'
-apply_recipe 'slim_rails'
-apply_recipe 'mini_racer'
-apply_recipe 'xray_rails'
-apply_recipe 'basic_user_dashboard'
+apply_recipe 'simple_form' # must run after bootstrap
+apply_recipe 'secure_headers'
+apply_recipe 'foreman'
 
 # template post hooks
 apply_recipe 'post_hooks'