+++ recipes/bundler_audit.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 add_to_gem_group :development do
-  gem_latest_version 'bundler-audit'
+  gem_latest_version 'bundler-audit', require: false
 end