+++ recipes/mini_racer.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-gem_latest_version 'mini_racer', platform: :ruby