+++ recipes/overcommit.rb
@@ -5,4 +5,5 @@ gem_install 'overcommit'
 after_bundle do
   copy_file 'templates/.overcommit.yml', '.overcommit.yml', force: true
   run 'overcommit --install'
+  run 'overcommit --sign'
 end