+++ recipes/simple_form.rb
@@ -5,6 +5,6 @@ return unless yes_install_gem? 'simple_form'
 gem_latest_version 'simple_form'
 
 after_bundle do
-  options = '--bootstrap' if gem_installed?('bootstrap', true)
+  options = '--bootstrap' if yarn_package_installed?('bootstrap')
   generate 'simple_form:install', options
 end