+++ template.rb
@@ -3,6 +3,7 @@
 require_relative 'lib/template'
 require_relative 'lib/console'
 require_relative 'lib/gem'
+require_relative 'lib/yarn'
 require_relative 'lib/system'
 
 def source_paths